+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `python-mode-major` no longer binds individual `send` commands.
   `inferior-python-mode` now enables `corfu`.
 - Package `wal-func` was renamed to `wal-useful`.
+- Package `wal-external` was renamed to `wal-package`.
 
 ### Removed